# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Österreichische EEG (Energiegemeinschaft) Abrechnungsplattform — REST API",
        "title": "eegabrechnung API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "eegabrechnung Support",
            "email": "support@eeg.at"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8101",
    "basePath": "/api/v1",
    "paths": {
...
        "/eegs/{eegID}/energy/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns raw energy readings aggregated per member for a date range. Used for CSV/XLSX export of member energy data. Both from and to are required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Berichte"
                ],
                "summary": "Raw per-member energy export",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, inclusive, Vienna time)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, inclusive, Vienna time)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.MemberStat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/eegs/{eegID}/energy/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns EEG-wide aggregated energy readings (consumption, generation, Restbedarf, Resteinspeisung) bucketed by the requested granularity. Date boundaries are interpreted in Vienna local time (Europe/Vienna).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Berichte"
                ],
                "summary": "EEG-wide energy summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, inclusive, Vienna time)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, inclusive, Vienna time)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "year",
                            "month",
                            "day",
                            "15min"
                        ],
                        "type": "string",
                        "description": "Time bucket size: year, month, day, 15min (default: month)",
                        "name": "granularity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.EnergySummaryRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/eegs/{eegID}/reports/annual": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates an XLSX annual report with two sheets: member list at the reference date and per-member energy + billing totals for the date range.",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Berichte"
                ],
                "summary": "Annual report (XLSX download)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Member reference date (YYYY-MM-DD) — members active on this date are listed",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Period start (YYYY-MM-DD) for energy and billing",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Period end (YYYY-MM-DD, inclusive) for energy and billing",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/eegs/{eegID}/reports/energy": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns aggregated energy and financial data grouped by calendar month for the given year. Each row contains consumption/generation kWh and revenue/payout totals.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Berichte"
                ],
                "summary": "Monthly energy report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Calendar year (default: current year)",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.MonthlyEnergyRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/eegs/{eegID}/reports/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns aggregated invoice totals and energy volumes per member over an optional date range. Used for the members breakdown report.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Berichte"
                ],
                "summary": "Per-member energy and invoice statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, inclusive)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, inclusive)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.MemberStat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.EnergySummaryRow": {
            "type": "object",
            "properties": {
                "period": {
                    "type": "string"
                },
                "wh_community": {
                    "description": "EEG share fed in (GENERATION wh_community)",
                    "type": "number"
                },
                "wh_restbedarf": {
                    "description": "grid demand = wh_total_consumption - wh_self",
                    "type": "number"
                },
                "wh_resteinspeisung": {
                    "description": "grid export  = wh_total_generation - wh_community",
                    "type": "number"
                },
                "wh_self": {
                    "description": "EEG share consumed (CONSUMPTION wh_self)",
                    "type": "number"
                },
                "wh_total_consumption": {
                    "description": "total consumption (CONSUMPTION wh_total)",
                    "type": "number"
                },
                "wh_total_generation": {
                    "description": "total generation (GENERATION wh_total)",
                    "type": "number"
                }
            }
        },
...
        "domain.MemberStat": {
            "type": "object",
            "properties": {
                "consumption_kwh": {
                    "description": "EEG share consumed (wh_self)",
                    "type": "number"
                },
                "consumption_total_kwh": {
                    "description": "total consumption (wh_total); 0 for billed mode",
                    "type": "number"
                },
                "generation_kwh": {
                    "description": "EEG share fed in (wh_community)",
                    "type": "number"
                },
                "generation_total_kwh": {
                    "description": "total generation (wh_total); 0 for billed mode",
                    "type": "number"
                },
                "invoice_count": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "domain.MonthlyEnergyRow": {
            "type": "object",
            "properties": {
                "consumption_kwh": {
                    "type": "number"
                },
                "generation_kwh": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                },
                "payouts": {
                    "description": "absolute sum of negative invoice totals (producer credits)",
                    "type": "number"
                },
                "revenue": {
                    "description": "sum of positive invoice totals (consumer charges incl. VAT)",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}